Restructured main if block in getEntryPointCodeFromShaderCache; Post-rebase cleanup
